Changed integer types for pointbasedstats

+++ BKM/source/pointbasedstat.hpp
@@ -7,18 +7,18 @@ namespace BKM {
 
 	class PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 	public:
 		PointBasedStat() : max(0), current(0), bonus(0) {};
-		PointBasedStat(uint32_t max) : max(max), current(max), bonus(0) {};
-		PointBasedStat(uint32_t max, uint32_t current) : max(max), current(current), bonus(0) {};
-		PointBasedStat(uint32_t max, uint32_t current, uint32_t bonus) : max(max), current(current), bonus(bonus) {};
+		PointBasedStat(uint64_t max) : max(max), current(max), bonus(0) {};
+		PointBasedStat(uint64_t max, uint64_t current) : max(max), current(current), bonus(0) {};
+		PointBasedStat(uint64_t max, uint64_t current, uint64_t bonus) : max(max), current(current), bonus(bonus) {};
 
 		// For adding points to 'current' field. ex increasing health vs maxhealth.
-		virtual void addPoints(uint32_t points) {
-			uint32_t limit = getMaxPoints();
+		virtual void addPoints(uint64_t points) {
+			uint64_t limit = getMaxPoints();
 			if (current = ~limit) {
 				if ((current + points) >= limit) {
 					current = limit;
@@ -30,7 +30,7 @@ namespace BKM {
 		};
 
 		// for removing points from the 'current' field.
-		virtual void removePoints(uint32_t points) {
+		virtual void removePoints(uint64_t points) {
 			if (current > 0 and points <= current) {
 				current = current - points;
 			}
@@ -40,13 +40,13 @@ namespace BKM {
 		};
 
 		// for adding to the 'max' field. ex increasing max health vs health.
-		virtual void addMaxPoints(uint32_t points) {
+		virtual void addMaxPoints(uint64_t points) {
 			max += points;
 		};
 
 		// for removing points from the 'max' field. 
-		virtual void removeMaxPoints(uint32_t points) {
-			uint32_t diff = max - points;
+		virtual void removeMaxPoints(uint64_t points) {
+			uint64_t diff = max - points;
 			if (current > diff) {
 				current = diff;
 			}
@@ -54,7 +54,7 @@ namespace BKM {
 		};
 
 		// for handling the addition of bonus points, we can add to both 'current' and 'bonus' fields with true for second parameter.
-		virtual void addBonus(uint32_t points, bool increaseCurrentPoints = false) {
+		virtual void addBonus(uint64_t points, bool increaseCurrentPoints = false) {
 			if (increaseCurrentPoints) {
 				current += points;
 			}
@@ -62,8 +62,8 @@ namespace BKM {
 		};
 
 		// for handling the removal of bonus points.
-		virtual void removeBonus(uint32_t points) {
-			uint32_t diff = bonus - points;
+		virtual void removeBonus(uint64_t points) {
+			uint64_t diff = bonus - points;
 			if (current > diff) {
 				current = diff;
 			}
@@ -73,108 +73,108 @@ namespace BKM {
 
 		// Get and Set Methods //
 
-		virtual const uint32_t getCurrentPoints() {
+		virtual const uint64_t getCurrentPoints() {
 			return current;
 		};
 
-		virtual void setCurrentPoints(uint32_t points) {
+		virtual void setCurrentPoints(uint64_t points) {
 			current = points;
 		};
 
-		virtual const uint32_t getMaxPoints() {
+		virtual const uint64_t getMaxPoints() {
 			return bonus + max;
 		};
 
-		virtual void setMaxPoints(uint32_t points) {
+		virtual void setMaxPoints(uint64_t points) {
 			max = points;
 		};
 
-		virtual const uint32_t getBonusPoints() {
+		virtual const uint64_t getBonusPoints() {
 			return bonus;
 		};
 
-		virtual void setBonusPoints(uint32_t points) {
+		virtual void setBonusPoints(uint64_t points) {
 			bonus = points;
 		};
 	};
 
 	class HealthPoints : public PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 
 	public:
 		HealthPoints() : PointBasedStat() {};
-		HealthPoints(uint32_t max) : PointBasedStat(max) {};
-		HealthPoints(uint32_t max, uint32_t current) : PointBasedStat(max, current) {};
-		HealthPoints(uint32_t max, uint32_t current, uint32_t bonus) : PointBasedStat(max, current, bonus) {};
+		HealthPoints(uint64_t max) : PointBasedStat(max) {};
+		HealthPoints(uint64_t max, uint64_t current) : PointBasedStat(max, current) {};
+		HealthPoints(uint64_t max, uint64_t current, uint64_t bonus) : PointBasedStat(max, current, bonus) {};
 	};
 
 
 	class ManaPoints : public PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 
 	public:
 		ManaPoints() : PointBasedStat() {};
-		ManaPoints(uint32_t max) : PointBasedStat(max) {};
-		ManaPoints(uint32_t max, uint32_t current) : PointBasedStat(max, current) {};
-		ManaPoints(uint32_t max, uint32_t current, uint32_t bonus) : PointBasedStat(max, current, bonus) {};
+		ManaPoints(uint64_t max) : PointBasedStat(max) {};
+		ManaPoints(uint64_t max, uint64_t current) : PointBasedStat(max, current) {};
+		ManaPoints(uint64_t max, uint64_t current, uint64_t bonus) : PointBasedStat(max, current, bonus) {};
 	};
 
 	class SoulPoints : public PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 
 	public:
 		SoulPoints() : PointBasedStat() {};
-		SoulPoints(uint32_t max) : PointBasedStat(max) {};
-		SoulPoints(uint32_t max, uint32_t current) : PointBasedStat(max, current) {};
-		SoulPoints(uint32_t max, uint32_t current, uint32_t bonus) : PointBasedStat(max, current, bonus) {};
+		SoulPoints(uint64_t max) : PointBasedStat(max) {};
+		SoulPoints(uint64_t max, uint64_t current) : PointBasedStat(max, current) {};
+		SoulPoints(uint64_t max, uint64_t current, uint64_t bonus) : PointBasedStat(max, current, bonus) {};
 	};
 
 	class EnergyPoints : public PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 
 	public:
 		EnergyPoints() : PointBasedStat() {};
-		EnergyPoints(uint32_t max) : PointBasedStat(max) {};
-		EnergyPoints(uint32_t max, uint32_t current) : PointBasedStat(max, current) {};
-		EnergyPoints(uint32_t max, uint32_t current, uint32_t bonus) : PointBasedStat(max, current, bonus) {};
+		EnergyPoints(uint64_t max) : PointBasedStat(max) {};
+		EnergyPoints(uint64_t max, uint64_t current) : PointBasedStat(max, current) {};
+		EnergyPoints(uint64_t max, uint64_t current, uint64_t bonus) : PointBasedStat(max, current, bonus) {};
 	};
 
 	class SpeedPoints : public PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 
 	public:
 		SpeedPoints() : PointBasedStat() {};
-		SpeedPoints(uint32_t max) : PointBasedStat(max) {};
-		SpeedPoints(uint32_t max, uint32_t current) : PointBasedStat(max, current) {};
-		SpeedPoints(uint32_t max, uint32_t current, uint32_t bonus) : PointBasedStat(max, current, bonus) {};
+		SpeedPoints(uint64_t max) : PointBasedStat(max) {};
+		SpeedPoints(uint64_t max, uint64_t current) : PointBasedStat(max, current) {};
+		SpeedPoints(uint64_t max, uint64_t current, uint64_t bonus) : PointBasedStat(max, current, bonus) {};
 	};
 
 	class ExperiencePoints : public PointBasedStat {
 	protected:
-		uint32_t max = 0;
-		uint32_t current = 0;
-		uint32_t bonus = 0;
+		uint64_t max = 0;
+		uint64_t current = 0;
+		uint64_t bonus = 0;
 
 	public:
 		ExperiencePoints() : PointBasedStat() {};
-		ExperiencePoints(uint32_t max) : PointBasedStat(max) {};
-		ExperiencePoints(uint32_t max, uint32_t current) : PointBasedStat(max, current) {};
-		ExperiencePoints(uint32_t max, uint32_t current, uint32_t bonus) : PointBasedStat(max, current, bonus) {};
+		ExperiencePoints(uint64_t max) : PointBasedStat(max) {};
+		ExperiencePoints(uint64_t max, uint64_t current) : PointBasedStat(max, current) {};
+		ExperiencePoints(uint64_t max, uint64_t current, uint64_t bonus) : PointBasedStat(max, current, bonus) {};
 	};
 
 } // BKM
